[#8] feature: base app renamed
