+++ leeteasy/models/challenge.py
